+++ src/repl.h
@@ -1,0 +1,8 @@
+#ifndef ILMA_REPL_H
+#define ILMA_REPL_H
+
+#include "evaluator.h"
+
+void repl_start(void);
+
+#endif